# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/users/me": {
      "get": {
        "operationId": "UsersController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/courses": {
      "get": {
        "operationId": "CoursesController_findPublished",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "courses"
        ]
      },
      "post": {
        "operationId": "CoursesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "courses"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/courses/{id}": {
      "get": {
        "operationId": "CoursesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "courses"
        ]
      },
      "patch": {
        "operationId": "CoursesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "courses"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "CoursesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "courses"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Online Ta'lim API",
    "description": "Videokurslar platformasi backend API",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "ali_valiyev",
            "description": "Faqat harf, raqam va _"
          },
          "password": {
            "type": "string",
            "example": "StrongPass123",
            "minLength": 8
          },
          "email": {
            "type": "string",
            "example": "student@example.com"
          },
          "firstName": {
            "type": "string",
            "example": "Ali"
          },
          "lastName": {
            "type": "string",
            "example": "Valiyev"
          },
          "city": {
            "type": "string",
            "example": "Toshkent"
          },
          "country": {
            "type": "string",
            "example": "uz",
            "description": "Davlat kodi"
          },
          "birthDay": {
            "type": "number",
            "example": 15,
            "minimum": 1,
            "maximum": 31
          },
          "birthMonth": {
            "type": "number",
            "example": 6,
            "minimum": 1,
            "maximum": 12
          },
          "birthYear": {
            "type": "number",
            "example": 2000,
            "minimum": 1940
          },
          "phone": {
            "type": "string",
            "example": "+998901234567"
          },
          "universityId": {
            "type": "string",
            "description": "Universitet ID (Hanafiy fiqh fakulteti talabalari uchun)"
          },
          "whatsapp": {
            "type": "string",
            "example": "+998901234567"
          },
          "nationality": {
            "type": "string",
            "example": "O'zbekiston",
            "description": "Fuqarolik"
          }
        },
        "required": [
          "username",
          "password",
          "email",
          "firstName",
          "lastName",
          "birthDay",
          "birthMonth",
          "birthYear",
          "nationality"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "ali_valiyev"
          },
          "password": {
            "type": "string",
            "example": "StrongPass123"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "CreateCourseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "JavaScript noldan"
          },
          "description": {
            "type": "string",
            "example": "Ushbu kursda JavaScript asoslarini o'rganasiz..."
          },
          "thumbnail": {
            "type": "string",
            "example": "https://cdn.example.com/thumb.jpg"
          },
          "price": {
            "type": "number",
            "example": 199000,
            "description": "Narx (so'mda). 0 = bepul"
          },
          "level": {
            "type": "string",
            "enum": [
              "BEGINNER",
              "INTERMEDIATE",
              "ADVANCED"
            ],
            "default": "BEGINNER"
          },
          "categoryId": {
            "type": "string",
            "description": "Kategoriya ID"
          }
        },
        "required": [
          "title",
          "description"
        ]
      },
      "UpdateCourseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "JavaScript noldan"
          },
          "description": {
            "type": "string",
            "example": "Ushbu kursda JavaScript asoslarini o'rganasiz..."
          },
          "thumbnail": {
            "type": "string",
            "example": "https://cdn.example.com/thumb.jpg"
          },
          "price": {
            "type": "number",
            "example": 199000,
            "description": "Narx (so'mda). 0 = bepul"
          },
          "level": {
            "type": "string",
            "enum": [
              "BEGINNER",
              "INTERMEDIATE",
              "ADVANCED"
            ],
            "default": "BEGINNER"
          },
          "categoryId": {
            "type": "string",
            "description": "Kategoriya ID"
          },
          "isPublished": {
            "type": "boolean",
            "description": "Kursni e'lon qilish"
          }
        }
      }
    }
  }
}