# src/courses/dto/update-course.dto.ts
import { ApiPropertyOptional, PartialType } from '@nestjs/swagger';
import { IsBoolean, IsOptional } from 'class-validator';
import { CreateCourseDto } from './create-course.dto';

export class UpdateCourseDto extends PartialType(CreateCourseDto) {
  @ApiPropertyOptional({ description: 'Kursni e\'lon qilish' })
  @IsOptional()
  @IsBoolean()
  isPublished?: boolean;
}